Add idle nodes in between open active nodes
When nodes have idle time in between them (when waiting for a parallel task),
we can add idle tasks to fix glitches in the display.

In addition, the todo list is sorted before being sent to the public-facing
server; this prevents issues in misaligned active tasks.

Finally, removed an extraneous import to walmartJSON in app.py.