Found bug with sorting on server
This is because of using node_list.sort vs sorted(node_list).

Although they should behave the same, they do not. Locally, it has different results,
but on the server, they act the same. Still not sure why, but we are working on it.

Might have to implement a custom sort.